Backend/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
ansi2html==1.8.0
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
category-encoders==2.5.1.post0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==3.0.0a11
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
fastapi==0.87.0
fastjsonschema==2.16.2
Flask==2.1.3
Flask-Cors==3.0.10
flatbuffers==22.11.23
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
idna==3.4
imbalanced-learn==0.9.1
importlib-metadata==5.1.0
ipywidgets==8.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
jupyter-dash==0.4.2
jupyterlab-widgets==3.0.3
kaleido==0.2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lightgbm==3.3.3
llvmlite==0.38.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.3
nbformat==5.7.0
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.3
pandas==1.4.4
patsy==0.5.3
Pillow==9.3.0
plotly==5.11.0
plotly-resampler==0.8.3
pmdarima==1.8.5
protobuf==3.19.6
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pydeck==0.8.0
Pympler==1.0.1
pyod==1.0.6
pyrsistent==0.19.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.4
rich==12.6.0
rsa==4.9
schemdraw==0.15
scikit-learn==1.1.3
scikit-plot==0.3.7
scipy==1.8.1
seaborn==0.12.1
semver==2.13.0
sktime==0.13.4
smmap==5.0.0
sniffio==1.3.0
starlette==0.21.0
statsmodels==0.13.5
streamlit==1.15.1
tbats==1.1.1
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tqdm==4.64.1
trace-updater==0.0.9
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
uvicorn==0.20.0
validators==0.20.0
watchdog==2.1.9
websockets==10.4
Werkzeug==2.1.2
widgetsnbextension==4.0.3
wincertstore==0.2
wrapt==1.14.1
yarl==1.8.2
yellowbrick==1.5
zipp==3.10.0
